add Makefile for AsyncNet in linux platform

+++ asyncnet/system.h
@@ -46,14 +46,14 @@
 #include <vector>
 #include <iostream>
 
-#include "imembase.h"
-#include "imemdata.h"
-#include "inetbase.h"
-#include "inetcode.h"
-#include "inetnot.h"
-#include "ineturl.h"
-#include "iposix.h"
-#include "itoolbox.h"
+#include "src/imembase.h"
+#include "src/imemdata.h"
+#include "src/inetbase.h"
+#include "src/inetcode.h"
+#include "src/inetnot.h"
+#include "src/ineturl.h"
+#include "src/iposix.h"
+#include "src/itoolbox.h"
 
 #include <stdio.h>
 #include <assert.h>
@@ -2032,7 +2032,7 @@ static inline void StringStrip(std::string &str, const char *seps) {
 		if (skip == 0) 
 			break;
 	}
-	for (p2 = str.size() - 1; p2 >= 0 && p2 >= p1; p2--) {
+	for (p2 = str.size() - 1; (ssize_t)p2 >= 0 && p2 >= p1; p2--) {
 		char ch = str[p2];
 		int skip = 0;
 		for (i = 0; seps[i]; i++) {
